% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cap_precip.R
\name{cap_precip}
\alias{cap_precip}
\title{Cap and Trade and Threshold and Trade Precipiation}
\usage{
cap_precip(path, ncrit.up = 20, ncrit.lo = NA, query.recursive = FALSE)
}
\arguments{
\item{path}{a character specifying the path to the location of climate files. Can be folder or just one file}

\item{ncrit.up}{a numeric value which specifies the maximum precipitation (unit according to input file). If \code{NA}, cap not used}

\item{ncrit.lo}{a numeric value which specifies the minimum precipitation (unit according to input file), default \code{NA}, meaning the cap is not used.}

\item{query.recursive}{if a folder is given, option exists to look for sub-folders, default \code{FALSE}.}
}
\value{
nothing, saves and overwrites weather files.
}
\description{
A function to produce lower and upper caps on the precipitation in XN5 weather files.
Overwrites former files. If a path is provided, and non-weather files are stored as .csv, it will not work.
Then, the explicit paths to the weather files need to be provided explicitly.
}
\examples{
\dontrun{
path <-  "./"
cap_precip(path, ncrit.up = 20, ncrit.lo = .5)
}

}
\author{
Tobias KD Weber, \email{tobias.weber@uni-hohenheim.de}
}
