% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse.R
\name{collapse}
\alias{collapse}
\title{Collapse a vector to a string}
\usage{
collapse(x, sep = ";")
}
\arguments{
\item{x}{a 1 column or 1 row data.frame or data.table, or a vector with n elements.}

\item{sep}{the seperator of the collapsed elements}
}
\value{
character vector of collapsed elements seperated by sep.
}
\description{
Intended for internal use.
}
\details{
Used to setup the cfg.ini, daisy_miner_nitrogen.ini, and leachn.ini
}
\examples{
\dontrun{
result <- collapse(c("f", "g"))
}
}
\author{
Tobias KD Weber , \email{tobias.weber@uni-hohenheim.de}
}
