% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_cfgClimateFile.R
\name{create_cfgClimateFile}
\alias{create_cfgClimateFile}
\title{Calculate Annual Average Temperature and Annual Monthly Temperature Amplitude}
\usage{
create_cfgClimateFile(
  path,
  query.write = FALSE,
  path_file_out = "cfgClimate.IN"
)
}
\arguments{
\item{path}{path to folder with csvs only climate files OR individual files}

\item{query.write}{defaults to \code{FALSE}. If TRUE a result file will be written , path_file_out = "cfgClimate.IN" a character specifying the path to the location of climate files. Can be}

\item{path_file_out}{name of the file}
}
\value{
nothing, saves and overwrites weather files.
}
\description{
The function is intended for use on the climate files in a given folder, to calculate the
\emph{AveYearTemp} and \emph{MonthTempAmp} for the cfg files. The output is stored in a file.
}
\examples{
\dontrun{
path <- c("./100_485 .csv", "./60_518 .csv")
path <- c("./100_485 .csv")
path <- c("./")
create_cfgClimateFile(path, query.write = TRUE)
}
}
\author{
Tobias KD Weber, \email{tobias.weber@uni-hohenheim.de}

Michelle Viswanathan, \email{michelle.viswanathan@uni-hohenheim.de}
}
