% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/croprotationDef.R
\name{croprotationDef}
\alias{croprotationDef}
\title{Setup crop_rotations.ini for XN5, based on DEFAULT management information}
\usage{
croprotationDef(x, start = "2014-08-1", end = "2019-08-1")
}
\arguments{
\item{x}{a character vector. Characters specify varieties, which should be incorporated
into the crop rotation. Varieties supported can be seen in the Details section.
The character vector will be recylced, until the end date, specified by the end arguments.}

\item{start}{character specifying the starting date of the crop rotation; format "yyyy-mm-dd": Has to be before the first sowing date corresponding to the first crop.}

\item{end}{character specifying the end date of the crop rotation; "yyyy-mm-dd": Set after to the last harvest date of the last crop.}
}
\value{
A list with three elements vector which can be used to produce the crop_rotation.ini
\item{start}{The date of the start of the crop rotation}
\item{end}{The date of the end of the crop rotation}
\item{CR}{The data.frame of the crop rotation to be written to crop_rotation.ini}
}
\description{
Setup crop_rotations.ini for XN5, based on DEFAULT management information
}
\details{
Accepted varieties are "Kartoffel", "Wintergerste", "SilageKRG", "Winterweizen","Sommergerste", "Winterrape", "zuckerruebe1", "Ackersenf"
}
\examples{
# example 1
}
\author{
Tobias KD Weber , \email{tobias.weber@uni-hohenheim.de}
}
