% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_CR.R
\docType{data}
\name{data_CR}
\alias{data_CR}
\title{crop rotations}
\format{
An object of class \code{"data.frame"}
}
\usage{
data(data_CR)
}
\description{
Data from the 42 soil water monitoring stations, the 6 FOR 1695 stations, and some text book rotations
}
\examples{
data(data_CR)
str(data_CR)
}
\author{
Tobias KD Weber , \email{tobias.weber@uni-hohenheim.de}
}
\keyword{datasets}
