% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fertilisationDef.R
\name{fertilisationDef}
\alias{fertilisationDef}
\title{Setup fertilisation.ini for XN5, based on DEFAULT fertilisation information and a prescribed
crop rotion dataframe from croprotationDef()}
\usage{
fertilisationDef(x)
}
\arguments{
\item{x}{the crop rotation data.frame from croprotationDef()}
}
\value{
\item{fertilisation}{A dataframe with information on fertilisation for a given crop.}
}
\description{
Setup fertilisation.ini for XN5, based on DEFAULT fertilisation information and a prescribed
crop rotion dataframe from croprotationDef()
}
\details{
Accepted varieties are "Kartoffel", "Wintergerste","SilageKRG" ,"Winterweizen","Sommergerste","Winterrape","zuckerruebe1","Ackersenf"
}
\examples{
# example 1
}
