% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cfgClimate.R
\name{get_cfgClimate}
\alias{get_cfgClimate}
\title{Calculate Annual average Temperature and Annual Monthly Temperature Amplitude}
\usage{
get_cfgClimate(x)
}
\arguments{
\item{x}{a list with at least the following 4 element
\itemize{
 \item{count$weather_NAMES}{The nams of the weather data files. Several can be provided }
 \item{sim_start}{Start date of the simulation Date class, format "YYYY-MM-DD"}
 \item{sim_end}{End date of the simulation Date class, format "YYYY-MM-DD"}
 \item{weather$data}{weather data from the weather.csv files}
}}
}
\value{
a list with 2 elements
\itemize{
 \item{\code{AveYearTemp}} {List with Average Annual Temperature for each climate file.}
 \item{\code{MonthTempAmp}}{List with Annual Monthly Temperature Amplitude}
}
}
\description{
Based on the weather data stored in settings, the Annual Average Temperature is calculated for the time period of complete years only.
There is no adjustment over the years (climate change), and it is not possible to do single years in this ewq
The function can be used to automatically deduce the maximum possible simulation window, given different weather files.
}
\examples{
\dontrun{
# no sensible example available. Provide x.
}
}
\author{
Tobias KD Weber, \email{tobias.weber@uni-hohenheim.de}

Michelle Viswanathan, \email{michelle.viswanathan@uni-hohenheim.de}
}
