% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_columnNames_xno.R
\name{get_columnNames_xno}
\alias{get_columnNames_xno}
\title{Extract column names from xno output file.}
\usage{
get_columnNames_xno(file.path = "my.xnos")
}
\arguments{
\item{file.path}{character vector specifying the at least one files path.}
}
\value{
vector with character of output state variables.
}
\description{
A simple function to quickly access the available output column names in the xno files.
}
\examples{
\dontrun{
# initialise list

# assign paths:
colnam <- get_columnNames_xno("my.xno")
}
}
\author{
Tobias KD Weber , \email{tobias.weber@uni-hohenheim.de}
}
