% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_simStartEnd.R
\name{get_simStartEnd}
\alias{get_simStartEnd}
\title{Recuperate the start and end time of a simulation based on available dates in weather files}
\usage{
get_simStartEnd(x, sep_arg = ",", query.head = TRUE)
}
\arguments{
\item{x}{a list with at least two elements
\itemize{
 \item{file_name}{The name of the weather data file. Several can be provided }
 \item{path}{The path of where the weather data file is stored. Either the same amount as provided in \code{file_name}
 or 1.}
}}

\item{sep_arg}{The column seperator in the weather file \code{DEFAULT = ,}.}

\item{query.head}{\code{TRUE/FALSE} If the weather file has a header or not. \code{DEFAULT = TRUE}.}
}
\value{
x List with time updated, crop specific DEFAULTs (fertilisation or tillage)
\itemize{
 \item{\code{file_name}}{ Character vector with the names of the weather data file. Several can be provided }
 \item{\code{path}} {Character vector with the names of the paths  where the weather data file is stored. Either the same amount as provided in \code{file_name}
  or 1.}
}
}
\description{
The weather file has to be in format required by Expert-N. Daily values are given in the format "YYYY-MM-DD"
The function can be used to automatically deduce the maximum possible simulation window, given different weather files.
}
\examples{
\dontrun{
}
}
\author{
Tobias KD Weber , \email{tobias.weber@uni-hohenheim.de}
}
