% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_XN5output_xno.R
\name{get_XN5output_xno}
\alias{get_XN5output_xno}
\title{setup XPN options line}
\usage{
get_XN5output_xno(mylocations, timestamps, query.return = FALSE)
}
\arguments{
\item{mylocations}{list
\itemize{
  \item{"my.xno.path"}{Character vector specifying the location of the result files. If a higher order folder is used, this will be included in the output. Possibility to make results unique.}
  \item{"my.output.path}{Character vector specifying the location of the output files. The files are ascii files and will be supplemented by \emph{.OUT}.}
}}

\item{timestamps}{character vector specifying the date(s) which are to be returned. Format "YYYY-MM-DD". No checks, careful. Most probably cause of errors, here.}

\item{query.return}{logical. FALSE is the default. If TRUE, a list is returned.}
}
\value{
If query.return == TRUE, returns list with number of elements as .xno files. Each element contains a data.table with the read timestamps.
}
\description{
A simple function that sets up the [options] block in an .xnp file. A DEFAULT setting is given, additional
moduls can be added by providing a list with named elements.
The name of the list element has to correspond to new line added.
}
\examples{
\dontrun{
# initialise list
mylocations <- list()

# assign paths:
# 1. location of xno.paths
mylocations$my.xno.path <- "C:/Projects/ExpertN/expertn513d/built/cfg/replicateECv10/"
# 2. location of output
mylocations$my.output.path <- "C:/Projects/ExpertN/myresults"

timestamps<- c("2009-08-02", "2026-09-03")
get_XN5output_xno(mylocations, timestamps)
}
}
\author{
Tobias KD Weber , \email{tobias.weber@uni-hohenheim.de}
}
