% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/header_xnweather.R
\docType{data}
\name{header_xnweather}
\alias{header_xnweather}
\title{header_xnweather}
\format{
An object of class \code{data.frame}
}
\usage{
data(header_xnweather)
}
\description{
This is a description
}
\examples{
data(header_xnweather)
str(header_xnweather)

}
\author{
Tobias KD Weber , \email{tobias.weber@uni-hohenheim.de}
}
\keyword{datasets}
