% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ini_cfgDef.R
\name{ini_cfgDef}
\alias{ini_cfgDef}
\title{setup cfg input}
\usage{
ini_cfgDef(
  simulation = NA,
  location = NA,
  climate = NA,
  preceding_crop = NA,
  soil,
  start_values,
  farm = NA
)
}
\arguments{
\item{simulation}{a vector with two named elements start and end, both characters, giving the dates (YYYY-MM-DD) of the simulation (should coincide with the weather file)}

\item{location}{a vector with named elements specifying location of the simulations, as well as the wind measurement height (wind_measure_height) and
the temperature measurement height (temp_measure_height)}

\item{climate}{a vector with named elements specifying the average annual temperature (AveYearTemp; from weather file) and monthly Temperature amplitude (MonthTempAmp; from weather file)}

\item{preceding_crop}{a vector with named elements giving details on the preceeding crop. Best left unchanged for long simulations}

\item{soil}{the list returned from the fucntion which determines the model setup of ini_cfgsoilDef}

\item{start_values}{the list returned from the function which determines the initial conditions ini_cfgicDef}

\item{farm}{a vector of named elements giving details which should be changed}
}
\value{
Setup the cfg.ini given several arguments
}
\description{
Soil parameterisation setup and returns a vector of characters for each line in the _cfg.ini file. Main inputs are i) soil hydraulic properties, and ii) initial conditions,
Through the arguments, the Default settings can be changed. arguments soil and start_values have to be provided
}
\examples{
\dontrun{
# no sensible example}
}
\author{
Tobias KD Weber , \email{tobias.weber@uni-hohenheim.de}
}
