% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_iniList.R
\name{merge_iniList}
\alias{merge_iniList}
\title{Convert cfg, leachn, daisy_miner_nitrogen template lists to data.frame}
\usage{
merge_iniList(x)
}
\arguments{
\item{x}{list, the modified cfg, leachn, daisy_miner_nitrogen template lists.}
}
\value{
Character vector of merged elements seperated by sep.
}
\description{
Intended for internal use.
}
\details{
Used to setup the variable which can be written to the _cfg.ini, _daisy_miner_nitrogen.ini, and _leachn.ini
}
\examples{
\dontrun{
}
}
\author{
Tobias KD Weber , \email{tobias.weber@uni-hohenheim.de}
}
