% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_cfgClimateFile.R
\docType{data}
\name{path}
\alias{path}
\title{Calculate Annual average Temperature and Annual Monthly Temperature Amplitude to Store in List}
\format{
An object of class \code{character} of length 1.
}
\usage{
path
}
\arguments{
\item{path}{a character specifying the path to the location of climate files. Can be folder or just one file.}
}
\value{
nothing, saves and overwrites weather files.
}
\description{
The function is intended for use on the climate files in a given folder, to calculate the
\emph{AveYearTemp} and \emph{MonthTempAmp} for the cfg files.
}
\examples{
\dontrun{
# no sensible example available. Provide x.
}
}
\author{
Tobias KD Weber, \email{tobias.weber@uni-hohenheim.de}

Michelle Viswanathan, \email{michelle.viswanathan@uni-hohenheim.de}
}
\keyword{datasets}
