% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_tpl.R
\name{read_tpl}
\alias{read_tpl}
\title{Read template files from the cfg_template folder in built}
\usage{
read_tpl(path = "./cfg_template", file.name = NA)
}
\arguments{
\item{path}{a character specifying the path to the template file location. Default to "./cfg_template".}

\item{file.name}{a character specifying the template file name.}
}
\value{
A named list with list element names the variable names (left of the =-sign),
and the elements being the template elements to the right of the =-sign.
These can be replaced.
}
\description{
The XN5 cfg_built folder contains template files to e.g. cfg.ini, leachn.ini, and
 daisy_miner_nitrogen.ini files. With this function, these files are read and made use of.
 As such, all comment lines in the template files (rows starting with #) are deleted.
}
\examples{
## No example possible/sensble
}
\author{
Tobias KD Weber, \email{tobias.weber@uni-hohenheim.de}
}
