% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schaaf_managDef.R
\name{schaaf_managDef}
\alias{schaaf_managDef}
\title{Setup schaaf_manag.ini for XN5, based on DEFAULT management information and a prescribed
crop rotion dataframe from croprotationDef()}
\usage{
schaaf_managDef(x)
}
\arguments{
\item{x}{the crop rotation data.frame from croprotationDef()}
}
\value{
\item{schaaf_manag}{A dataframe with information on the schaaf_manag for a given crop}.
}
\description{
Setup schaaf_manag.ini for XN5, based on DEFAULT management information and a prescribed
crop rotion dataframe from croprotationDef()
}
\details{
Accepted varieties are "Kartoffel", "Wintergerste","SilageKRG" ,"Winterweizen","Sommergerste","Winterrape","zuckerruebe1","Ackersenf"
}
\examples{
# example 1
}
