% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sel_GridElements.R
\name{sel_GridElements}
\alias{sel_GridElements}
\title{Pre-select grid elements)}
\usage{
sel_GridElements(data.vector, select.vector)
}
\arguments{
\item{data.vector}{character vector specifying the element names.}

\item{select.vector}{null, numbers or named elements of length > 1 to select from data.vector. These will be selected
using \code{sel$soilunit},\code{sel$weather},\code{sel$ptf},\code{sel$croprotation}. Character elements only sensible, if in data.vector character elements.}
}
\value{
returns a reduced data.vector, depending on select.vector
}
\description{
The function can be used to select individual elements from the stochastic 4D dimensions.
}
\examples{
\dontrun{
# Difficiult to concrete examples
# res <- sel_GridElements(data.vector     = sel$BK50$soilunits
#        , select.vector = NULL )

# res <- sel_GridElements(data.vector     = sel$BK50$soilunits
#        , select.vector =  1)

# res <- sel_GridElements(data.vector     = sel$BK50$soilunits
#        , select.vector = c(1,2) )

# res <- sel_GridElements(data.vector     = sel$BK50$soilunits
#        , select.vector = "g34" )

# res <- sel_GridElements(data.vector     = sel$BK50$soilunits
#        , select.vector = c("g34". "g35") )

}


}
