% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soilProfileDef.R
\name{soilProfileDef}
\alias{soilProfileDef}
\title{Calculate depth distributions}
\usage{
soilProfileDef(
  zdepth,
  layer_thickness = 5,
  pars = list(Adepth = 25, n = 2, fact = 1),
  plot.query = FALSE,
  rescale.query = FALSE
)
}
\arguments{
\item{zdepth}{depth of the soil profile [cm].}

\item{layer_thickness}{thickness of the simulation layers.}

\item{pars}{list of three named elements. Parameters determining the shape of the profile distribution.
#' \describe{
  \item{Adrepth}{depth of A horizon [cm] }
  \item{n}{sharpness of drop at Adepth [-]}
  \item{fact}{factor by which the entire profile (top = 1), is multiplied. For example by a humus content [-]}
}}

\item{plot.query}{query if a plot of the depth profile should be made. TRUE = plot is provided. FALSE, no plot.}

\item{rescale.query}{query if the matter content in the soil should be rescaled to 0 at the bottom. TRUE, rescales.}
}
\value{
Vector of calculated values ordered from top to bottom, in depth increments provided by \code{layer_thickness}
}
\description{
Synthetic distribution of carbon and nitrogen contents in soilscontent with depth
}
\details{
Used in the functions schaaf_managDef and in fertilisationDef
}
\examples{
\dontrun{result <- soilProfileDef(200, plot.query = TRUE, rescale.query = FALSE)}
}
\author{
Tobias KD Weber , \email{tobias.weber@uni-hohenheim.de}
}
