% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tpl_gecros.R
\docType{data}
\name{tpl_gecros}
\alias{tpl_gecros}
\title{Crop model parameters of gecros}
\format{
An object of class \code{"list"}. Each list element contains the \code{"readLines"} of the original parameter template file.
        The parameterisation are the standard parameterisation from FOR 1695 Regional Climate Change.
}
\usage{
data(tpl_gecros)
}
\description{
Crop model parameter files for 15 different calibrated varieties
}
\examples{
data(tpl_gecros)
str(tpl_gecros)
}
\author{
Tobias KD Weber , \email{tobias.weber@uni-hohenheim.de}
}
\keyword{datasets}
