% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_list.R
\name{update_list}
\alias{update_list}
\title{Update the DEFAULT settings of the xpi and xpn lists, fixed in various function.}
\usage{
update_list(x, opt.list)
}
\arguments{
\item{x}{a list with named elements passed to the function. Will update opt.list accordingls.}

\item{opt.list}{named list to be updated by x}
}
\value{
character vector, which can be concatenated to, e.g. the xpi and xpn definition files
}
\description{
Intended for internal use.
}
\details{
Used in definition functions
}
\examples{
\dontrun{
}
}
\author{
Tobias KD Weber , \email{tobias.weber@uni-hohenheim.de}
}
