% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_simStartEnd.R
\name{update_simStartEnd}
\alias{update_simStartEnd}
\title{Update the start and end simulation dates.}
\usage{
update_simStartEnd(x)
}
\arguments{
\item{x}{The list with named elements specifying the settings of the simulation experiment.}
}
\value{
x a list with updated elements.
}
\description{
The start and end dates of the simulation experiment can be given. If the climate data files
do not cover that period, these will be update. Similarly, if the climate data files contain data
beyond the simulation start and end date, the simulation experiment will be done for the time specified.
If no simulation start and end dates were specified, these will be set to the first and last timestamp
in the climate data, respectively.#'
}
\examples{
\dontrun{
}
}
\author{
Tobias KD Weber , \email{tobias.weber@uni-hohenheim.de}
}
