% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_time.R
\name{update_time}
\alias{update_time}
\title{Update application times of the DEFAULT fertiliser and tillage}
\usage{
update_time(x, DT)
}
\arguments{
\item{x}{a list returned from croprotationDef. If not tinkered with, there is a chornological order in the list which is highly (!) adivised.}

\item{DT}{(default) DT_min, DT_org, DT_manag which have to contain a column named "date" with the dates of activity as a characrter in "YYYY-MM-DD" format.}
}
\value{
List with time updated, crop specific DEFAULTs (fertilisation or tillage)
}
\description{
A simple function used to update DEFAULT dates based on the information given in croprotation, defined by the list returned from croprotationDef
}
\details{
Used in the functions schaaf_managDef and in fertilisationDef
}
\examples{
## example 1
\dontrun{
}
}
\author{
Tobias KD Weber , \email{tobias.weber@uni-hohenheim.de}
}
