% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_iniList.R
\name{write_iniList}
\alias{write_iniList}
\title{Convert and write to files the cfg, leachn, daisy_miner_nitrogen initialisations.}
\usage{
write_iniList(xlist, file_header, file_paths = NA, append.query = FALSE)
}
\arguments{
\item{xlist}{list with croprotations to be written to the file specified in \code{file_paths}.}

\item{file_header}{character, header of the .ini file, e.g. "[plant management]" in the croprotation.ini.}

\item{file_paths}{character vector with the file names corresponding to the crop rotations in \code{xlist}.}

\item{append.query}{\code{TRUE} or \code{FALSE}, specifying if the output should be appended to existing files.}
}
\value{
If \code{file_paths = NA} then the character vector is returned, which of \code{file_paths}
are provided, is saved to a file.
}
\description{
Intended for internal use.
}
\details{
This function is specific to the list of returned elements from \code{croprotationDef}
}
\examples{
\dontrun{
}
}
\author{
Tobias KD Weber , \email{tobias.weber@uni-hohenheim.de}
}
