% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xn3xn5_climate.R
\name{xn3xn5_climate}
\alias{xn3xn5_climate}
\title{get XN3 weather file climate info}
\usage{
xn3xn5_climate(xn3_xnw = NA, path_in = NA)
}
\arguments{
\item{xn3_xnw}{character xn3.xnw. Careful of paths}

\item{path_in}{directory of XN3 weather files. If xn3_xnw is \code{NA}, xnw files in (sub-)directories are non-recursively searched and converted.}
}
\description{
Get XN3 climate information (Marker 20000) for XN5 cfg input.
}
\examples{

}
\author{
Tobias KD Weber , \email{tobias.weber@uni-hohenheim.de}
}
