% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xn3xn5_dateFormat.R
\name{xn3xn5_dateFormat}
\alias{xn3xn5_dateFormat}
\title{convert XN3 date format to XN5 date format}
\usage{
xn3xn5_dateFormat(xn3_dates = 10100, cutoff = 79)
}
\arguments{
\item{xn3_dates}{single numeric value or vector of numerical values.}

\item{cutoff}{In XN3, only 2 digits are used for the Year. This makes it complicated. \code{cutoff} is the two digit value,
after which all values are thought of belonging to the 20th centure (i.e. 19xx). Otherwise, it is interpreted to be a year from 2000.}
}
\description{
Use to convert XN3 date format "ddmmyy" to XN5 "yyy-mm-dd"
}
\examples{
result <- xn3xn5_dateFormat(xn3_dates = 10100, cutoff = 79)
}
\author{
Tobias KD Weber , \email{tobias.weber@uni-hohenheim.de}
}
