% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xn3xn5_weather.R
\name{xn3xn5_weather}
\alias{xn3xn5_weather}
\title{convert XN3 weather files to XN5 weather files.}
\usage{
xn3xn5_weather(xn3_xnw = NA, print_to_file = TRUE, path_in = NA, path_out = NA)
}
\arguments{
\item{xn3_xnw}{character xn3.xnw. Careful of paths}

\item{print_to_file}{\code{TRUE} result is printed to XN5 file, \code{FALSE}, time series of the XN3 weather file are returned.}

\item{path_in}{directory of XN3 weather files. If xn3_xnw is \code{NA}, xnw files in (sub-)directories are non-recursively searched and converted.}

\item{path_out}{directory to store XN5 weather files. If \code{NA}, then directory specified in path_in is used, else, current directory.}
}
\description{
Use to convert XN3 weather files (.xnw) to XN5 weather files (.csv).
}
\examples{

}
\author{
Tobias KD Weber , \email{tobias.weber@uni-hohenheim.de}
}
