% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XN5setup-package.R
\docType{package}
\name{XN5setup-package}
\alias{XN5setup}
\alias{XN5setup-package}
\title{XN5setup: Setup semi-deterministic XN5 projects}
\description{
R package to set up gridded XN5 simulations. The core idea is to setup a z, x, y, tile grid (in XN5: grid, i, j, tile), where:
    \emph{z} is the vertical component, used as the grid layers (e.g. stochastic weather),
    \emph{x} is the spatial grid abscissa, e.g. spatial explicit properties,
    \emph{y} is the spatial grid ordinate, e.g. for stochastic soil properties, and
    \emph{tile} are the grid cell's sub-columns, e.g. crop rotations.
}
\author{
\strong{Maintainer}: Tobias KD WEBER \email{<tobias.weber@uni-hohenheim.de>} (\href{https://orcid.org/0000-0002-3448-5208}{ORCID})

}
\keyword{internal}
