% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xpn_modelActivate.R
\name{xpn_modelActivate}
\alias{xpn_modelActivate}
\title{Activate/Deactive individual cells from outside the XN5 gui}
\usage{
xpn_modelActivate(
  xpn_file_path,
  xpi_activate = "all",
  flip = FALSE,
  xpn_file_path_out = NULL
)
}
\arguments{
\item{xpn_file_path}{.xpn file in question, full path names ahas to be given.}

\item{xpi_activate}{character to specify the plant model. Supported: GECROS, GECROS_h, SPASS, CERES}

\item{flip}{character to specify the mineralisation model. Supported: DAISY and LEACHN.}

\item{xpn_file_path_out}{optional chaarcter to write a new xpn, default \code{NULL}, means overwriting of .xpn specified in \code{xpn_file_path}. Make sure to provide a full path, otherwise the new xpn will be written into the current working directory.}
}
\value{
nothing. The function overwrites the xpn if \code{xpn_file_path_out} is not specified-.
}
\description{
A simple function that enables to rapiud selection/deselection of simulated cells in the xpn.
}
\examples{

# xpn_modelActivate(xpn_file_path, xpi_activate = c("0_0_0_0.xpi", "0_1_0_0.xpi"), flip = TRUE)

}
\author{
Tobias KD Weber , \email{tobias.weber@uni-hohenheim.de}
}
