% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xpn_modelCut.R
\name{xpn_modelCut}
\alias{xpn_modelCut}
\title{Split the .xpn (i.e. Master) file into several smaller projects}
\usage{
xpn_modelCut(xpn_file_path, splits.no = NULL, chunks.len = NULL)
}
\arguments{
\item{xpn_file_path}{a list with at least the following 4 element}

\item{splits.no}{number of splits along the xpn.}

\item{chunks.len}{number of simulation units per xpn chunk}
}
\description{
The function enable to sperate the xpn grid into sub-elements, so that
}
\examples{
\dontrun{
# no sensible example available. Provide x.
}
}
\author{
Tobias KD Weber, \email{tobias.weber@uni-hohenheim.de}
}
