% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xpn_modelDef.R
\name{xpn_modelDef}
\alias{xpn_modelDef}
\title{setup XPN model line (the grid)}
\usage{
xpn_modelDef(x, PROJECT_NAME = NA)
}
\arguments{
\item{x}{a list with named elements. The names of the elements will be used to construct the grid layout in the .xpn.
All grid cells are activated by default. See \code{details}}

\item{PROJECT_NAME}{character specifying the name of the project. May be included in x, instead.}
}
\value{
The grid to be created in the .xnp file
}
\description{
Defaults to all active
A simple function that sets up the [model] grid = line in an .xnp file.
Default and only setup form is:
grid coordinate j ('x') for the soil units, e.g. BK50, BÜK200, BÜK1000,
grid coordinate i ('y') for simple stochastification, e.g. PTFs
grid stack number ('z') the number of weather scenarios.
tile element ('tile') for the crop rotations
}
\details{
x$weather_NAMES  the names of the different weather scenarios OR the integer number of grid layers
x$soil_unit_NAMES the names of the different soil units to be used OR the total number
x$stoch_NAMES the names of the different stochastic untis to be used (e.g. number of PTF) OR the total number
x$CR_NAMES the names of tiles

Alternative to the vector of names, the integer number of unique elements can be given in x.
}
\examples{
## example 1
x <- list()
x$PROJECT_NAME    <- "test_XN5setup"
x$soil_unit_NAMES <- 10L
x$stoch_NAMES     <- 3L
x$weather_NAMES   <- 2L
x$CR_NAMES        <- 3L
res <- xpn_modelDef(x)
}
\author{
Tobias KD Weber , \email{tobias.weber@uni-hohenheim.de}
}
