% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xpn_modulDef.R
\name{xpn_modulDef}
\alias{xpn_modulDef}
\title{Setup XPN module line}
\usage{
xpn_modulDef(x = NA)
}
\arguments{
\item{x}{a list with named elements. The names of the elements will be used as modules to be loaded.}
}
\value{
The libs specification in the .xnp file
}
\description{
A simple function that sets up the [modul] libs = line in an .xnp file. A DEFAULT setting is given, additional
moduls can be added by providing a list with named elements. The name should correspond to new module which should be additionally loaded
}
\examples{
\dontrun{
## example 1
libsreturn1 <- xpn_modulDef()
## example 2
libsini <- list()
libsini$testmodule <- 1
libsreturn2 <- xpn_modulDef(libsini)
}
}
\author{
Tobias KD Weber , \email{tobias.weber@uni-hohenheim.de}
}
