% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xpn_optionsDef.R
\name{xpn_optionsDef}
\alias{xpn_optionsDef}
\title{setup XPN options line}
\usage{
xpn_optionsDef(x = NA, query.plant.model = "gecros")
}
\arguments{
\item{x}{i) a list with named elements. The names of the elements will be used as the LHS and
the vector element will be the RHS of the equal sign in extending [options].
ii) The default options can be updated by named x elements equalling the [options] LHS}

\item{query.plant.model}{character to specify the plant model. Supported: GECROS, GECROS_h}
}
\value{
The options specification for the .xnp file
}
\description{
A simple function that sets up the [options] block in an .xnp file. A DEFAULT setting is given, additional
moduls can be added by providing a list with named elements.
The name of the list element has to correspond to new line added.
}
\examples{
# Default settings can be seen by executing the following line
opts <- xpn_optionsDef(x = NA)

}
\author{
Tobias KD Weber , \email{tobias.weber@uni-hohenheim.de}
}
