% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xpn_systemDef.R
\name{xpn_systemDef}
\alias{xpn_systemDef}
\title{setup XPN system line}
\usage{
xpn_systemDef(x = NA)
}
\arguments{
\item{x}{NA - Default values are used i) a list with named elements. The names of the elements will be used as the LHS and
the vector element will be the RHS of the equal sign in extending [system].
ii) The default system can be updated by named x elements equalling the [system] LHS}
}
\value{
The system specification for the .xnp file
}
\description{
A simple function that sets up the [system] block in an .xnp file. A DEFAULT setting is given, additional
moduls can be added by providing a list with named elements.
The name of the list element has to correspond to new line added.
}
\examples{
# Default settings can be seen by executing the following line
syts <- xpn_systemDef(x = NA)

}
\author{
Tobias KD Weber , \email{tobias.weber@uni-hohenheim.de}
}
